+++ src/Components/sidebar/docs/changelog.md
@@ -1,4 +1,4 @@
-# 15th July 2026
+15th July 2026
 - Built
 - Added icons
 - Restructured
@@ -7,11 +7,28 @@
 - Built dropdown
 - Used browser calculations
 
-# 16th July 2026
+16th July 2026
 - Built Variant
 - Added theme switching
 - Added theme variation
 
-# 17th July 2026
+17th July 2026
 - Added animations
 - Built active state indicator
+
+18th July 2026
+- Added MPA navigation
+- Made into web component
+- Restructured
+- Used Modules
+- Re-used
+- Added cross document transitions
+
+19th July 2026
+- Added SPA navigation
+- Added Routing
+- Built router function
+
+20th July 2026
+- Used imports
+- Added view transitions